fix : place CRUD with jwt authorization

+++ src/main/java/com/ssafy/luminous/place/service/PlaceService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -54,14 +55,23 @@ public class PlaceService {
            return placeRepository.save(placePostReqDto.toEntity(memberId));
     }
 
-    public void deletePlace(Long id) {
+    public void deletePlace(Long id, Long memberId) {
+        Optional<Place> place = placeRepository.findByIdAndMember_id(id,memberId);
+        if(place.isEmpty()){
+            System.out.println("사용자의 게시글이 아니에요~");
+            throw new RuntimeException("사용자 게시글이 아녜요");
+        }
         placeRepository.deleteById(id);
     }
 
-    public Place updatePlace(Long id, PlaceUpdateReqDto placeUpdateReqDto){
-        Place place = placeRepository.findById(id).orElseThrow();
-        place.update(placeUpdateReqDto);
-        return place;
+    public Place updatePlace(Long id, PlaceUpdateReqDto placeUpdateReqDto, Long memberId){
+        Optional<Place> place = placeRepository.findByIdAndMember_id(id,memberId);
+        if(place.isEmpty()){
+            System.out.println("사용자의 게시글이 아니에요~");
+            throw new RuntimeException("사용자 게시글이 아녜요");
+        }
+        place.get().update(placeUpdateReqDto);
+        return place.get();
 
     }
